patient_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
drf-nested-routers==0.93.4 